fix: Move model setup code to separate file
